# Protect against concurrent reads/writes to Context keyvalues