--- a/concurrency-tests/build.gradle
+++ b/concurrency-tests/build.gradle
@@ -5,7 +5,6 @@ plugins {
 dependencies {
     implementation project(":micrometer-core")
 //    implementation("io.micrometer:micrometer-core:1.12.4")
-    implementation project(":micrometer-test")
     implementation project(":micrometer-registry-prometheus")
     runtimeOnly(libs.logbackLatest)
 }