--- a/micrometer-commons/src/main/java/io/micrometer/common/KeyValues.java
+++ b/micrometer-commons/src/main/java/io/micrometer/common/KeyValues.java
@@ -371,12 +371,7 @@ public final class KeyValues implements Iterable<KeyValue> {
         }
         else if (keyValues instanceof Collection) {
             Collection<? extends KeyValue> keyValuesCollection = (Collection<? extends KeyValue>) keyValues;
-            try {
-                return toKeyValues(keyValuesCollection.toArray(EMPTY_KEY_VALUE_ARRAY));
-            }
-            catch (ArrayIndexOutOfBoundsException exception) {
-                return of(keyValues);
-            }
+            return toKeyValues(keyValuesCollection.toArray(EMPTY_KEY_VALUE_ARRAY));
         }
         else {
             return toKeyValues(StreamSupport.stream(keyValues.spliterator(), false).toArray(KeyValue[]::new));