See https://github.com/micrometer-metrics/micrometer/issues/4356 and https://github.com/micrometer-metrics/tracing/issues/898.