@shakuzen 2024-12-12
And sharing benchmark results from my machine (MBP M1). There was not much of a noticeable difference and at the level of nanoseconds, the run-to-run results can vary as much as the differences here.

With these changes using ConcurrentHashMap:

```
Benchmark                                                         Mode      Cnt       Score   Error  Units
ObservationKeyValuesBenchmark.contended                         sample  6816019     140.191 ± 0.406  ns/op
ObservationKeyValuesBenchmark.contended:contendedRead           sample  2979345     133.465 ± 0.694  ns/op
ObservationKeyValuesBenchmark.contended:contendedRead:p0.00     sample                  ≈ 0          ns/op
ObservationKeyValuesBenchmark.contended:contendedRead:p0.50     sample              125.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedRead:p0.90     sample              250.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedRead:p0.95     sample              250.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedRead:p0.99     sample              375.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedRead:p0.999    sample              459.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedRead:p0.9999   sample            14784.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedRead:p1.00     sample           153856.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedWrite          sample  3836674     145.414 ± 0.480  ns/op
ObservationKeyValuesBenchmark.contended:contendedWrite:p0.00    sample                  ≈ 0          ns/op
ObservationKeyValuesBenchmark.contended:contendedWrite:p0.50    sample              125.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedWrite:p0.90    sample              250.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedWrite:p0.95    sample              250.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedWrite:p0.99    sample              375.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedWrite:p0.999   sample              458.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedWrite:p0.9999  sample            12255.960          ns/op
ObservationKeyValuesBenchmark.contended:contendedWrite:p1.00    sample           116480.000          ns/op
ObservationKeyValuesBenchmark.contended:p0.00                   sample                  ≈ 0          ns/op
ObservationKeyValuesBenchmark.contended:p0.50                   sample              125.000          ns/op
ObservationKeyValuesBenchmark.contended:p0.90                   sample              250.000          ns/op
ObservationKeyValuesBenchmark.contended:p0.95                   sample              250.000          ns/op
ObservationKeyValuesBenchmark.contended:p0.99                   sample              375.000          ns/op
ObservationKeyValuesBenchmark.contended:p0.999                  sample              458.000          ns/op
ObservationKeyValuesBenchmark.contended:p0.9999                 sample            13696.000          ns/op
ObservationKeyValuesBenchmark.contended:p1.00                   sample           153856.000          ns/op
ObservationKeyValuesBenchmark.uncontendedRead                   sample  1634837      21.613 ± 0.825  ns/op
ObservationKeyValuesBenchmark.uncontendedRead:p0.00             sample                  ≈ 0          ns/op
ObservationKeyValuesBenchmark.uncontendedRead:p0.50             sample                  ≈ 0          ns/op
ObservationKeyValuesBenchmark.uncontendedRead:p0.90             sample               42.000          ns/op
ObservationKeyValuesBenchmark.uncontendedRead:p0.95             sample               42.000          ns/op
ObservationKeyValuesBenchmark.uncontendedRead:p0.99             sample               42.000          ns/op
ObservationKeyValuesBenchmark.uncontendedRead:p0.999            sample              166.000          ns/op
ObservationKeyValuesBenchmark.uncontendedRead:p0.9999           sample             1708.000          ns/op
ObservationKeyValuesBenchmark.uncontendedRead:p1.00             sample           380416.000          ns/op
ObservationKeyValuesBenchmark.uncontendedWrite                  sample  1401661      23.337 ± 0.350  ns/op
ObservationKeyValuesBenchmark.uncontendedWrite:p0.00            sample                  ≈ 0          ns/op
ObservationKeyValuesBenchmark.uncontendedWrite:p0.50            sample               41.000          ns/op
ObservationKeyValuesBenchmark.uncontendedWrite:p0.90            sample               42.000          ns/op
ObservationKeyValuesBenchmark.uncontendedWrite:p0.95            sample               42.000          ns/op
ObservationKeyValuesBenchmark.uncontendedWrite:p0.99            sample               42.000          ns/op
ObservationKeyValuesBenchmark.uncontendedWrite:p0.999           sample              166.000          ns/op
ObservationKeyValuesBenchmark.uncontendedWrite:p0.9999          sample             2083.335          ns/op
ObservationKeyValuesBenchmark.uncontendedWrite:p1.00            sample            73344.000          ns/op
```

Before these changes (1.14.1) using LinkedHashMap:

```
Benchmark                                                         Mode      Cnt        Score   Error  Units
ObservationKeyValuesBenchmark.contended                         sample  6777815      143.705 ± 1.565  ns/op
ObservationKeyValuesBenchmark.contended:contendedRead           sample  2938907      138.407 ± 1.226  ns/op
ObservationKeyValuesBenchmark.contended:contendedRead:p0.00     sample                   ≈ 0          ns/op
ObservationKeyValuesBenchmark.contended:contendedRead:p0.50     sample               125.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedRead:p0.90     sample               250.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedRead:p0.95     sample               250.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedRead:p0.99     sample               333.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedRead:p0.999    sample               417.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedRead:p0.9999   sample             19206.989          ns/op
ObservationKeyValuesBenchmark.contended:contendedRead:p1.00     sample            432128.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedWrite          sample  3838908      147.762 ± 2.598  ns/op
ObservationKeyValuesBenchmark.contended:contendedWrite:p0.00    sample                   ≈ 0          ns/op
ObservationKeyValuesBenchmark.contended:contendedWrite:p0.50    sample               125.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedWrite:p0.90    sample               250.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedWrite:p0.95    sample               250.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedWrite:p0.99    sample               333.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedWrite:p0.999   sample               417.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedWrite:p0.9999  sample             15248.000          ns/op
ObservationKeyValuesBenchmark.contended:contendedWrite:p1.00    sample           2068480.000          ns/op
ObservationKeyValuesBenchmark.contended:p0.00                   sample                   ≈ 0          ns/op
ObservationKeyValuesBenchmark.contended:p0.50                   sample               125.000          ns/op
ObservationKeyValuesBenchmark.contended:p0.90                   sample               250.000          ns/op
ObservationKeyValuesBenchmark.contended:p0.95                   sample               250.000          ns/op
ObservationKeyValuesBenchmark.contended:p0.99                   sample               333.000          ns/op
ObservationKeyValuesBenchmark.contended:p0.999                  sample               417.000          ns/op
ObservationKeyValuesBenchmark.contended:p0.9999                 sample             16512.000          ns/op
ObservationKeyValuesBenchmark.contended:p1.00                   sample           2068480.000          ns/op
ObservationKeyValuesBenchmark.uncontendedRead                   sample  1663755       18.553 ± 0.231  ns/op
ObservationKeyValuesBenchmark.uncontendedRead:p0.00             sample                   ≈ 0          ns/op
ObservationKeyValuesBenchmark.uncontendedRead:p0.50             sample                   ≈ 0          ns/op
ObservationKeyValuesBenchmark.uncontendedRead:p0.90             sample                42.000          ns/op
ObservationKeyValuesBenchmark.uncontendedRead:p0.95             sample                42.000          ns/op
ObservationKeyValuesBenchmark.uncontendedRead:p0.99             sample                42.000          ns/op
ObservationKeyValuesBenchmark.uncontendedRead:p0.999            sample               125.000          ns/op
ObservationKeyValuesBenchmark.uncontendedRead:p0.9999           sample              1400.225          ns/op
ObservationKeyValuesBenchmark.uncontendedRead:p1.00             sample             69760.000          ns/op
ObservationKeyValuesBenchmark.uncontendedWrite                  sample  1401784       23.264 ± 0.479  ns/op
ObservationKeyValuesBenchmark.uncontendedWrite:p0.00            sample                   ≈ 0          ns/op
ObservationKeyValuesBenchmark.uncontendedWrite:p0.50            sample                41.000          ns/op
ObservationKeyValuesBenchmark.uncontendedWrite:p0.90            sample                42.000          ns/op
ObservationKeyValuesBenchmark.uncontendedWrite:p0.95            sample                42.000          ns/op
ObservationKeyValuesBenchmark.uncontendedWrite:p0.99            sample                42.000          ns/op
ObservationKeyValuesBenchmark.uncontendedWrite:p0.999           sample               166.000          ns/op
ObservationKeyValuesBenchmark.uncontendedWrite:p0.9999          sample              1958.000          ns/op
ObservationKeyValuesBenchmark.uncontendedWrite:p1.00            sample            132608.000          ns/op

```